+++ app/mpd_client/client.h
@@ -48,6 +48,7 @@ namespace MpdClient {
     void updateDb();
     QByteArray albumArt(const QString& filepath);
     QByteArray readPicture(const QString& filepath);
+    bool renamePlaylist(const QString &old_name, const QString &new_name);
 
   signals:
     void connected(const QUrl &url);