+++ app/playlists_ui/playlistscontextmenu.h
@@ -28,6 +28,7 @@ namespace PlaylistsUi {
     void removed(const QModelIndex &index);
     void playlistChanged(const std::shared_ptr<Playlist::Playlist> pl);
     void loadPlaylistFiles(const QModelIndex &index, const QStringList &filepaths);
+    void renamed(const QString &old_name, const QString &new_name);
 
   private:
     ProxyFilterModel *model;